synapse-registry/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Limiter==3.5.0
SQLAlchemy==2.0.21
PyJWT==2.8.0
Werkzeug==2.3.7
python-dotenv==1.0.0
boto3==1.28.45
pytest==7.4.2
pytest-cov==4.1.0
